made an order what is domain and what is service/infra
