+++ backend/modules/data/domain/exceptions.py
@@ -5,6 +5,6 @@ class FileTypeNotSupportedErrod(Exception):
 
     
 class FileNotFoundError(Exception):
-    def __init__(self, file_format: str):
-        self.file_format = file_format
+    def __init__(self, file_path: str):
+        self.file_path = file_path
         super().__init__(self)
